requirements.txt
openai
python-dotenv # For API key management 